api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pika==1.3.2
redis==5.0.1
pydantic>=2.0.0
python-dotenv==1.0.0
python-multipart==0.0.6
authlib==1.2.1
python-jose==3.3.0
aiofiles==24.1.0
itsdangerous==2.2.0
httpx==0.25.0
bcrypt==4.1.2
email-validator==2.1.0
stripe==12.2.0
